fix: verify and close the review's sixteen open findings
Every unverified finding from the multi-agent review was adversarially
re-checked against current code: fifteen confirmed and fixed, one refuted
(the paste-back guard already lowercases both sides). Highlights: the
advisor pre-flight now reserves the whole generation budget, the busy
inference slot is an error the store re-arms on instead of a fake empty
answer, the Bonferroni family counts the turns arm it gates, sweep apply
re-scopes instead of deleting, open_external validates schemes, the
capability grant is trimmed to what the webview uses, launch-at-login
failures surface and revert the switch, Discover joins Savers cache-only
so a primary tab never phones GitHub unprompted, and the app copy is
free of em dashes. Clippy is at zero across the workspace.

+++ crates/piggy-core/tests/advisor_tests.rs
@@ -100,7 +100,7 @@ fn every_catalog_model_runs_on_8gb() {
 #[test]
 fn budget_reserves_room_for_the_rest_of_the_machine() {
     // Never plan around more than 60% of RAM, and never less than 3 GB reserved.
-    assert_eq!(budget(8 * GB), 8 * GB - (8 * GB as u64) * 2 / 5);
+    assert_eq!(budget(8 * GB), 8 * GB - (8 * GB) * 2 / 5);
     assert!(budget(4 * GB) < 2 * GB);
     assert!(budget(64 * GB) < 40 * GB);
 }
